% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_api.R
\name{bp_advance_time}
\alias{bp_advance_time}
\title{Advance Time}
\usage{
bp_advance_time(state, n_ticks = 1L)
}
\arguments{
\item{state}{Program state.}

\item{n_ticks}{Number of ticks to advance.}
}
\value{
Updated program state.
}
\description{
Advance simulation time by integer ticks.
}
