% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_api.R
\name{bp_latest_model_id}
\alias{bp_latest_model_id}
\title{Get Latest Model ID}
\usage{
bp_latest_model_id(state)
}
\arguments{
\item{state}{Program state.}
}
\value{
Character scalar model id, or `NULL`.
}
\description{
Return the id of the most recently trained model in `state$gs_models`.
}
