% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_print.R
\name{bp_state_overview}
\alias{bp_state_overview}
\title{State Overview}
\usage{
bp_state_overview(state)
}
\arguments{
\item{state}{A `bp_state` object.}
}
\value{
Named list with summary statistics.
}
\description{
Build a compact overview of a `bp_state` object.
}
