\name{BreedingProgramSimulator-package}
\alias{BreedingProgramSimulator}
\alias{BreedingProgramSimulator-package}
\title{BreedingProgramSimulator: Stateful AlphaSimR Breeding Program Architecture}
\description{
Package for building stateful, event-driven breeding program simulators on top of AlphaSimR.
}
\docType{package}
