% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_api.R
\name{close_sources}
\alias{close_sources}
\title{Close Source Cohorts}
\usage{
close_sources(state, source)
}
\arguments{
\item{state}{Program state.}

\item{source}{Source bundle from [get_ready_pop()].}
}
\value{
Updated program state.
}
\description{
Mark all source cohorts in a source bundle as inactive.
}
