% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_api.R
\name{run_genotyping}
\alias{run_genotyping}
\title{Run Genotyping}
\usage{
run_genotyping(state, cfg)
}
\arguments{
\item{state}{Program state.}

\item{cfg}{Genotyping configuration list.}
}
\value{
Updated program state.
}
\description{
Schedule/log genotyping events by cohort and chip, with genotyping costs.
}
