% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_wrappers.R
\name{run_multienv_trial}
\alias{run_multienv_trial}
\title{Run Example Multi-Environment Trial}
\usage{
run_multienv_trial(state)
}
\arguments{
\item{state}{Program state.}
}
\value{
Updated program state.
}
\description{
Readable template function for a multi-environment phenotyping stage.
}
